refactor: pipeline page model refs
