+++ source/javascripts/pages/PipelinesPage/PipelinesPage.mocks.ts
@@ -1,4 +1,4 @@
-import { BitriseYml } from '../../models/BitriseYml';
+import { BitriseYml } from '@/core/models/BitriseYml';
 
 export const mockYml: BitriseYml = {
   format_version: '13',
